test-client/requirements.txt
pytest
httpx
pyzmq